rename rider list subfolder
