+++ lib/widgets/pages/home_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:weforza/generated/l10n.dart';
-import 'package:weforza/widgets/pages/member_list/rider_list_page.dart';
 import 'package:weforza/widgets/pages/ride_list/ride_list_page.dart';
+import 'package:weforza/widgets/pages/rider_list/rider_list_page.dart';
 import 'package:weforza/widgets/pages/settings/settings_page.dart';
 import 'package:weforza/widgets/platform/platform_aware_widget.dart';
